+++ python/mxnet/__init__.py
@@ -8,8 +8,7 @@ from .base import MXNetError
 from . import base
 from . import contrib
 from . import ndarray
-from . import sparse_ndarray
-from . import ndarray_utils
+from . import ndarray as nd
 from . import name
 # use mx.sym as short for symbol
 from . import symbol as sym
@@ -18,10 +17,6 @@ from . import symbol_doc
 from . import io
 from . import recordio
 from . import operator
-# use mx.nd as short for mx.ndarray
-from . import ndarray as nd
-from . import sparse_ndarray as sparse_nd
-from . import ndarray_utils as nd_utils
 # use mx.rnd as short for mx.random
 from . import random as rnd
 from . import random
