+++ python/mxnet/ndarray/sparse_ndarray.py
@@ -15,31 +15,32 @@ import sys as _sys
 
 # import operator
 import numpy as np
-import mxnet as mx
-from .base import _LIB, numeric_types
-from .base import c_array, mx_real_t
-from .base import mx_uint, NDArrayHandle, check_call
-from .context import Context
-from . import _ndarray_internal as _internal
+from ..base import _LIB, numeric_types
+from ..base import c_array, mx_real_t
+from ..base import mx_uint, NDArrayHandle, check_call
+from ..context import Context
+from . import _internal
 from . import ndarray
 from .ndarray import _DTYPE_NP_TO_MX, _DTYPE_MX_TO_NP
 from .ndarray import _STORAGE_TYPE_STR_TO_ID
-from .ndarray import NDArray, _storage_type
+from .ndarray import NDArray, _storage_type, _zeros_ndarray
+from . import cast_storage
+from . import slice as nd_slice
 
 # Use different verison of SymbolBase
 # When possible, use cython to speedup part of computation.
 # pylint: disable=unused-import
 try:
     if int(_os.environ.get("MXNET_ENABLE_CYTHON", True)) == 0:
-        from ._ctypes.ndarray import NDArrayBase, _set_ndarray_class
+        from .._ctypes.ndarray import NDArrayBase, _set_ndarray_class
     elif _sys.version_info >= (3, 0):
-        from ._cy3.ndarray import NDArrayBase, _set_ndarray_class
+        from .._cy3.ndarray import NDArrayBase, _set_ndarray_class
     else:
-        from ._cy2.ndarray import NDArrayBase, _set_ndarray_class
+        from .._cy2.ndarray import NDArrayBase, _set_ndarray_class
 except ImportError:
     if int(_os.environ.get("MXNET_ENFORCE_CYTHON", False)) != 0:
         raise ImportError("Cython Module cannot be loaded but MXNET_ENFORCE_CYTHON=1")
-    from ._ctypes.ndarray import NDArrayBase, _set_ndarray_class
+    from .._ctypes.ndarray import NDArrayBase, _set_ndarray_class
 
 # pylint: enable=unused-import
 _STORAGE_AUX_TYPES = {
@@ -48,7 +49,7 @@ _STORAGE_AUX_TYPES = {
 }
 
 
-def _new_alloc_handle(storage_type, shape, ctx, delay_alloc, dtype, aux_types, aux_shapes=None):
+def _new_alloc_handle(stype, shape, ctx, delay_alloc, dtype, aux_types, aux_shapes=None):
     """Return a new handle with specified storage type, shape, dtype and context.
 
     Empty handle is only used to hold results
@@ -65,7 +66,7 @@ def _new_alloc_handle(storage_type, shape, ctx, delay_alloc, dtype, aux_types, a
     aux_shapes = sum(aux_shapes, ())
     num_aux = mx_uint(len(aux_types))
     check_call(_LIB.MXNDArrayCreateSparseEx(
-        ctypes.c_int(int(_STORAGE_TYPE_STR_TO_ID[storage_type])),
+        ctypes.c_int(int(_STORAGE_TYPE_STR_TO_ID[stype])),
         c_array(mx_uint, shape),
         mx_uint(len(shape)),
         ctypes.c_int(ctx.device_typeid),
@@ -114,13 +115,13 @@ class SparseNDArray(NDArray):
 
         Examples
         --------
-        >>> src = mx.sparse_nd.row_sparse(data, indices, (3,3))
+        >>> src = mx.nd.row_sparse([[1, 0, 2], [4, 5, 6]], [0, 2], (3,3))
         >>> src.asnumpy()
         array([[ 1.,  0.,  2.],
                [ 0.,  0.,  0.],
                [ 4.,  5.,  6.]], dtype=float32)
         >>> # assign SparseNDArray with same storage type
-        >>> x = mx.sparse_nd.zeros('row_sparse', (3,3))
+        >>> x = mx.nd.zeros('row_sparse', (3,3))
         >>> x[:] = src
         >>> x.asnumpy()
         array([[ 1.,  0.,  2.],
@@ -168,6 +169,7 @@ class SparseNDArray(NDArray):
 
         Examples
         --------
+        >>> x = mx.nd.zeros((2, 3), stype='row_sparse')
         >>> x[:] = mx.nd.arange(0,6).reshape((2,3))
         >>> x.asnumpy()
         array([[ 0.,  1.,  2.],
@@ -175,7 +177,7 @@ class SparseNDArray(NDArray):
         >>> x[1:2].asnumpy()
         array([[ 3.,  4.,  5.]], dtype=float32)
         """
-        stype = self.storage_type
+        stype = self.stype
         if stype != 'csr':
             raise Exception("__getitem__ for " + str(stype) + " not implemented yet")
         if isinstance(key, int):
@@ -186,7 +188,7 @@ class SparseNDArray(NDArray):
             if key.start is not None or key.stop is not None:
                 begin = key.start if key.start else 0
                 end = key.stop if key.stop else self.shape[0]
-                return ndarray.slice(self, begin=begin, end=end)
+                return nd_slice(self, begin=begin, end=end)
             else:
                 return self
         if isinstance(key, tuple):
@@ -217,7 +219,7 @@ class SparseNDArray(NDArray):
         return _DTYPE_MX_TO_NP[aux_type.value]
 
     @property
-    def values(self):
+    def data(self):
         """The values array of the SparseNDArray. This is a read-only view of the values array.
         They reveal internal implementation details and should be used with care.
 
@@ -228,12 +230,11 @@ class SparseNDArray(NDArray):
         """
         return self._data()
 
-
     @property
     def _num_aux(self):
         ''' The number of aux data used to help store the sparse ndarray.
         '''
-        return len(_STORAGE_AUX_TYPES[self.storage_type])
+        return len(_STORAGE_AUX_TYPES[self.stype])
 
     @property
     # pylint: disable= invalid-name, undefined-variable
@@ -241,7 +242,7 @@ class SparseNDArray(NDArray):
         raise Exception('Transpose is not supported for SparseNDArray.')
 
     @property
-    def aux_types(self):
+    def _aux_types(self):
         """The data types of the aux data for the SparseNDArray.
         """
         aux_types = []
@@ -264,13 +265,13 @@ class SparseNDArray(NDArray):
             The type of the returned array.
         Examples
         --------
-        >>> x = mx.sparse_nd.zeros('row_sparse', (2,3), dtype='float32')
+        >>> x = mx.nd.zeros('row_sparse', (2,3), dtype='float32')
         >>> y = x.astype('int32')
         >>> y.dtype
         <type 'numpy.int32'>
         """
-        res = mx.nd.zeros(shape=self.shape, ctx=self.context,
-                          dtype=dtype, storage_type=self.storage_type)
+        res = _zeros_sparse_ndarray(shape=self.shape, ctx=self.context,
+                                    dtype=dtype, stype=self.stype)
         self.copyto(res)
         return res
 
@@ -301,8 +302,8 @@ class SparseNDArray(NDArray):
                 return
             return _internal._copyto(self, out=other)
         elif isinstance(other, Context):
-            hret = _ndarray_cls(_new_alloc_handle(self.storage_type, self.shape, other,
-                                                  True, self.dtype, self.aux_types))
+            hret = _ndarray_cls(_new_alloc_handle(self.stype, self.shape, other,
+                                                  True, self.dtype, self._aux_types))
             return _internal._copyto(self, out=hret)
         else:
             raise TypeError('copyto does not support type ' + str(type(other)))
@@ -333,7 +334,18 @@ class CSRNDArray(SparseNDArray):
     row i are stored in indices[indptr[i]:indptr[i+1]] and their corresponding values are stored
     in values[indptr[i]:indptr[i+1]].
 
+    Example
+    -------
+    >>> a = mx.nd.array([[0, 1, 0], [2, 0, 0], [0, 0, 0], [0, 0, 3]])
+    >>> a = a._to_csr()
+    >>> a.indices.asnumpy()
+    array([1, 0, 2])
+    >>> a.indptr.asnumpy()
+    array([0, 1, 2, 2, 3])
+    >>> a.data.asnumpy()
+    array([ 1.,  2.,  3.], dtype=float32)
     """
+
     def __reduce__(self):
         return CSRNDArray, (None,), super(CSRNDArray, self).__getstate__()
 
@@ -376,6 +388,17 @@ class RowSparseNDArray(SparseNDArray):
 
     RowSparseNDArray is used principally in the definition of gradients for operations
     that have sparse gradients (e.g. SparseEmbedding).
+
+    Examples
+    --------
+    >>> import mxnet as mx
+    >>> dense = mx.nd.array([[1,2],[0,0],[3,0],[0,0]])
+    >>> rsp = dense._to_rsp()
+    >>> rsp.indices.asnumpy()
+    array([0, 2], dtype=int32)
+    >>> rsp.data.asnumpy()
+    array([[ 1.,  2.],
+           [ 3.,  0.]], dtype=float32)
     """
     def __reduce__(self):
         return RowSparseNDArray, (None,), super(RowSparseNDArray, self).__getstate__()
@@ -406,12 +429,12 @@ def _prepare_src_array(src, dtype, default_dtype):
     return src, dtype
 
 
-def csr(values, indptr, indices, shape, ctx=None, dtype=None, indptr_type=None, indices_type=None):
+def csr(data, indptr, indices, shape, ctx=None, dtype=None, indptr_type=None, indices_type=None):
     """Creates a 2D array with compressed sparse row format.
 
     Parameters
     ----------
-    values: array_like
+    data: array_like
         An object exposing the array interface, with shape [nnz], where D0 is the number of
         non-zero entries.
     indptr: array_like
@@ -419,9 +442,9 @@ def csr(values, indptr, indices, shape, ctx=None, dtype=None, indptr_type=None, 
         should always be zero.
     indices: array_like
         An object exposing the array interface, with shape [nnz].
-    ctx : Context, optional
+    ctx: Context, optional
         Device context (default is the current default context).
-    dtype : str or numpy.dtype, optional
+    dtype: str or numpy.dtype, optional
         The data type of the output array. The default dtype is ``values.dtype``
         if `values` is an `NDArray`, `float32` otherwise.
     indptr_type: str or numpy.dtype, optional
@@ -435,13 +458,23 @@ def csr(values, indptr, indices, shape, ctx=None, dtype=None, indptr_type=None, 
     -------
     CSRNDArray
         A `CSRNDArray` with the `csr` storage representation.
+
+    Example
+    -------
+    >>> import mxnet as mx
+    >>> a = mx.nd.csr([1, 2, 3], [0, 1, 2, 2, 3], [1, 0, 2], (4, 3))
+    >>> a.asnumpy()
+    array([[ 0.,  1.,  0.],
+           [ 2.,  0.,  0.],
+           [ 0.,  0.,  0.],
+           [ 0.,  0.,  3.]], dtype=float32)
     """
     storage_type = 'csr'
     # context
     if ctx is None:
         ctx = Context.default_ctx
     # prepare src array and types
-    values, dtype = _prepare_src_array(values, dtype, mx_real_t)
+    data, dtype = _prepare_src_array(data, dtype, mx_real_t)
     indptr, indptr_type = _prepare_src_array(indptr, indptr_type,
                                              _STORAGE_AUX_TYPES[storage_type][0])
     indices, indices_type = _prepare_src_array(indices, indices_type,
@@ -451,17 +484,17 @@ def csr(values, indptr, indices, shape, ctx=None, dtype=None, indptr_type=None, 
     assert('int64' in str(indices_type)), "expected int64 for indices"
     # verify shapes
     aux_shapes = [indptr.shape, indices.shape]
-    assert(values.ndim == 1)
+    assert(data.ndim == 1)
     assert(indptr.ndim == 1)
     assert(indices.ndim == 1)
     assert(len(shape) == 2)
     result = CSRNDArray(_new_alloc_handle(storage_type, shape, ctx, False, dtype,
                                           [indptr_type, indices_type], aux_shapes))
-    # assign indptr, indices and values
-    values_ref = result._data(True)
+    # assign indptr, indices and data
+    data_ref = result._data(True)
     indptr_ref = result._aux_data(0, True)
     indices_ref = result._aux_data(1, True)
-    values_ref[:] = values
+    data_ref[:] = data
     indptr_ref[:] = indptr
     indices_ref[:] = indices
     return result
@@ -490,6 +523,17 @@ def row_sparse(values, indices, shape, ctx=None, dtype=None, indices_type=None):
     -------
     RowSparseNDArray
         An `RowSparseNDArray` with the `row_sparse` storage representation.
+
+    Example
+    -------
+    >>> a = mx.nd.row_sparse([[1, 2], [3, 4]], [1, 4], (6, 2))
+    >>> a.asnumpy()
+    array([[ 0.,  0.],
+           [ 1.,  2.],
+           [ 0.,  0.],
+           [ 0.,  0.],
+           [ 3.,  4.],
+           [ 0.,  0.]], dtype=float32)
     """
     storage_type = 'row_sparse'
     # context
@@ -522,11 +566,12 @@ def todense(source):
     NDArray
         The dense array with default storage
     """
-    return ndarray.cast_storage(source, storage_type='default')
+    return cast_storage(source, stype='default')
 
 
-def _ndarray_cls(handle, writable=True):
-    stype = _storage_type(handle)
+def _ndarray_cls(handle, writable=True, stype=None):
+    if stype is None:
+        stype = _storage_type(handle)
     if stype == 'default':
         return NDArray(handle, writable=writable)
     elif stype == 'csr':
@@ -538,3 +583,46 @@ def _ndarray_cls(handle, writable=True):
 
 
 _set_ndarray_class(_ndarray_cls)
+
+
+def _zeros_sparse_ndarray(stype, shape, ctx=None, dtype=None, aux_types=None, **kwargs):
+    """Return a new array of given shape and type, filled with zeros.
+
+    Parameters
+    ----------
+    shape : int or tuple of int
+        The shape of the empty array
+    stype: string
+        The storage type of the empty array, such as 'row_sparse', 'csr', etc
+    ctx : Context, optional
+        An optional device context (default is the current default context)
+    dtype : str or numpy.dtype, optional
+        An optional value type (default is `float32`)
+    aux_types: list of numpy.dtype, optional
+        An optional type for the aux data for SparseNDArray (default values depends
+        on the storage type)
+
+    Returns
+    -------
+    SparseNDArray
+        A created array
+    Examples
+    --------
+    >>> mx.nd.zeros('csr', (1,2), mx.gpu(0))
+    <SparseNDArray 1x2 @gpu(0)>
+    >>> mx.nd.zeros('row_sparse', (1,2), mx.gpu(0), 'float16').asnumpy()
+    array([[ 0.,  0.]], dtype=float16)
+    """
+    if stype == 'default':
+        return _zeros_ndarray(shape, ctx=ctx, dtype=dtype, **kwargs)
+    if ctx is None:
+        ctx = Context.default_ctx
+    dtype = mx_real_t if dtype is None else dtype
+    if aux_types is None:
+        if stype == 'row_sparse' or stype == 'csr':
+            aux_types = _STORAGE_AUX_TYPES[stype]
+        else:
+            raise Exception("unknown storage type")
+    assert(len(aux_types) == len(_STORAGE_AUX_TYPES[stype]))
+    out = _ndarray_cls(_new_alloc_handle(stype, shape, ctx, True, dtype, aux_types))
+    return _internal._zeros(shape=shape, ctx=ctx, dtype=dtype, out=out, **kwargs)
